Install pinentry to allow gpg-agent interaction
This is my first time modifying a flatpak, and I don't know if this is
the best way to do this, but it seems to work! :) I got inspiration
from the org.gnome.Evolution flatpak. If it looks like something you
want, but would like me to change it in some way first, let me know.

I found the following issues which may allow this to be done better in
future:

Portal for GPG encryption/decryption
https://github.com/flatpak/xdg-desktop-portal/issues/178

gpg-agent socket
https://github.com/flatpak/flatpak/issues/2301

To make use of this modification, you can install the "pinentry" emacs
package (https://elpa.gnu.org/packages/pinentry.html), and add the
following to your `.emacs` configuration:

```elisp
(pinentry-start)
```