Add some support for SDK extensions
This fixes issue #31 by initially launching a wrapper script that
first adjusts our environment, and then launches the actual Emacs
executable.

The wrapper script is taken from the vscode-flatpak-wrapper
repository. Despite the name, this utility is suitable for other IDEs
as well.

Note: In this specific version, the wrapper submodule is taken from
my personal fork. This is intended to be pointed at the module
upstream once its pull request has been merged.